+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".BookingActivity"
+            android:name=".BookingDetailActivity"
             android:exported="false" />
         <activity
             android:name=".BookingCompleteActivity"
