+++ app/src/main/java/woowacourse/movie/booking/complete/BookingCompleteContract.kt
@@ -1,6 +1,7 @@
-package woowacourse.movie.model
+package woowacourse.movie.booking.complete
 
 import android.os.Bundle
+import woowacourse.movie.model.BookingResult
 
 interface BookingCompleteContract {
     interface Presenter {
