+++ app/src/main/java/woowacourse/movie/booking/detail/BookingResultUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui
+package woowacourse.movie.booking.detail
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
