+++ app/src/main/java/woowacourse/movie/model/BookingResult.kt
