+++ app/src/main/res/layout/activity_booking.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BookingActivity">
+    tools:context=".BookingDetailActivity">
 
     <ScrollView
         android:layout_width="match_parent"
