+++ app/src/test/java/woowacourse/movie/model/BookingResultTest.kt
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
+import woowacourse.movie.model.BookingResult
 import java.time.LocalDate
 import java.time.LocalTime
